+++ examples/move_control/drag_drop_control.ico
