+++ examples/move_control/small.ico
